refactor(test): move integration tests inline and reduce test_helpers visibility
Move the 7 `backend/tests/` files into `#[cfg(test)] mod tests` blocks
in the source modules they exercise, merge shared support helpers into
`src/test_helpers.rs`, and change the module from `pub` to crate-private.

Made-with: Cursor
